encapsulated header and fixed redux
