+++ src/Header/Left/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import HeaderLeft from './header-left';
+
+export default HeaderLeft;
