+++ src/Header/Left/styles.js
@@ -1,0 +1,26 @@
+const headerLeft_ = {
+  marginLeft: '8px',
+};
+
+const styles = {
+  image: {
+    color: 'darkgray',
+    fontSize: '21px',
+  },
+  ngine: {
+    color: 'gray',
+    fontSize: '21px',
+  },
+  title: {
+    ...headerLeft_,
+    display: 'inline-block',
+    verticalAlign: 'top',
+  },
+  panafold: {
+    ...headerLeft_,
+    height: '24px',
+    width: 'auto',
+  },
+};
+
+export default styles;
