+++ src/index.js
@@ -3,10 +3,13 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
+import injectTapEventPlugin from 'react-tap-event-plugin';
 import { createStore } from 'redux';
 
 import App from './App';
 import reducers from './reducers';
+
+injectTapEventPlugin();
 
 const preloadedState = window.__PRELOADED_STATE;
 delete window.__PRELOADED_STATE;
