+++ src/reducers/index.js
@@ -1,9 +1,10 @@
 import { combineReducers } from 'redux';
 
-import headerActions from './header-actions';
+import { searchOpen, uploadOpen } from './header-actions';
 import renderLocation from './render-location';
 
 export default combineReducers({
-  headerActions,
   renderLocation,
+  searchOpen,
+  uploadOpen,
 });
